requirements.txt
   langchain
   wikipedia
   langchain-community
   langchain-openai
   langchain-anthropic
   langchain-llama
   python-dotenv
   pydantic
   langchain_google_genai
   duckduckgo-search